Improved entry fields, basic widget layout, and added background
Made proper temporary text functions; temporary text that prompts to enter a value disappears when focused, and reappears when unfocussed if you have typed no text. A frame was made in the main program class as a background. Placed widgets in roughly the appropriate spots (temporary and will be improved upon later).

Co-Authored-By: GauravS2507 <135093120+GauravS2507@users.noreply.github.com>